torch_learn/pyproject.toml
[project]
name = "torch-learn"
version = "0.1.0"
description = "Comprehensive PyTorch tutorials from basics to advanced computer vision"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "gensim>=4.4.0",
    "pandas>=2.0.0",
    "transformers>=4.30.0",
]

[[tool.uv.index]]
url="https://pypi.tuna.tsinghua.edu.cn/simple"
default=true
